feat: create stack
